Re-organizing files, adding documentation
