+++ client/components/happychat/functional.js
