+++ client/lib/functional/test/index.js
@@ -1,0 +1,4 @@
+
+describe( 'functional', () => {
+	describe( 'not ready' );
+} );
