+++ client/state/action-types.js
@@ -10,18 +10,6 @@
  *  - https://wpcalypso.wordpress.com/devdocs/docs/our-approach-to-data.md#actions
  */
 
-// TODO: Move to correct place when ready for support chat
-
-export const HAPPYCHAT_CONNECT = 'HAPPYCHAT_CONNECT';
-export const HAPPYCHAT_CONNECTING = 'HAPPYCHAT_CONNECTING';
-export const HAPPYCHAT_CONNECTED = 'HAPPYCHAT_CONNECTED';
-export const HAPPYCHAT_CLOSING = 'HAPPYCHAT_CLOSING';
-export const HAPPYCHAT_SET_MESSAGE = 'HAPPYCHAT_SET_MESSAGE';
-export const HAPPYCHAT_RECEIVE_EVENT = 'HAPPYCHAT_RECEIVE_EVENT';
-export const HAPPYCHAT_SET_AUTOSCROLL = 'HAPPYCHAT_SET_AUTOSCROLL';
-export const HAPPYCHAT_OPEN = 'HAPPYCHAT_OPEN';
-export const HAPPYCHAT_OPEN_URL = 'HAPPYCHAT_OPEN_URL';
-
 export const ANALYTICS_EVENT_RECORD = 'ANALYTICS_EVENT_RECORD';
 export const ANALYTICS_MULTI_TRACK = 'ANALYTICS_MULTI_TRACK';
 export const ANALYTICS_PAGE_VIEW_RECORD = 'ANALYTICS_PAGE_VIEW_RECORD';
@@ -92,6 +80,15 @@ export const GUIDED_TRANSFER_STATUS_RECEIVE = 'GUIDED_TRANSFER_STATUS_RECEIVE';
 export const GUIDED_TRANSFER_STATUS_REQUEST = 'GUIDED_TRANSFER_STATUS_REQUEST';
 export const GUIDED_TRANSFER_STATUS_REQUEST_FAILURE = 'GUIDED_TRANSFER_STATUS_REQUEST_FAILURE';
 export const GUIDED_TRANSFER_STATUS_REQUEST_SUCCESS = 'GUIDED_TRANSFER_STATUS_REQUEST_SUCCESS';
+export const HAPPYCHAT_CONNECT = 'HAPPYCHAT_CONNECT';
+export const HAPPYCHAT_CONNECTING = 'HAPPYCHAT_CONNECTING';
+export const HAPPYCHAT_CONNECTED = 'HAPPYCHAT_CONNECTED';
+export const HAPPYCHAT_CLOSING = 'HAPPYCHAT_CLOSING';
+export const HAPPYCHAT_SET_MESSAGE = 'HAPPYCHAT_SET_MESSAGE';
+export const HAPPYCHAT_RECEIVE_EVENT = 'HAPPYCHAT_RECEIVE_EVENT';
+export const HAPPYCHAT_SET_AUTOSCROLL = 'HAPPYCHAT_SET_AUTOSCROLL';
+export const HAPPYCHAT_OPEN = 'HAPPYCHAT_OPEN';
+export const HAPPYCHAT_OPEN_URL = 'HAPPYCHAT_OPEN_URL';
 export const IMAGE_EDITOR_CROP = 'IMAGE_EDITOR_CROP';
 export const IMAGE_EDITOR_FLIP = 'IMAGE_EDITOR_FLIP';
 export const IMAGE_EDITOR_ROTATE_COUNTERCLOCKWISE = 'IMAGE_EDITOR_ROTATE_COUNTERCLOCKWISE';
